Move reusable code to isolated modules and apply formatting - no more global config
