+++ src/wam2layers/analysis/visualization.py
@@ -1,3 +1,4 @@
+import logging
 from pathlib import Path
 
 import click
@@ -9,9 +10,6 @@ from cmocean import cm
 from wam2layers.config import Config
 from wam2layers.preprocessing.shared import get_grid_info
 from wam2layers.tracking.backtrack import input_path, output_path
-
-import logging
-
 from wam2layers.utils import load_region
 
 logger = logging.getLogger(__name__)
@@ -91,7 +89,11 @@ def _plot_evap(config, ax):
 
     # Make figure
     e_track.plot(
-        ax=ax, vmin=0, robust=True, cmap=cm.rain, cbar_kwargs=dict(fraction=0.05, shrink=0.5)
+        ax=ax,
+        vmin=0,
+        robust=True,
+        cmap=cm.rain,
+        cbar_kwargs=dict(fraction=0.05, shrink=0.5),
     )
     e_track.plot.contour(ax=ax, levels=[0.1, 1], colors=["lightgrey", "grey"])
     ax.set_title("Moisture source of extreme precip [mm]", loc="left")
