+++ src/wam2layers/config.py
@@ -1,10 +1,10 @@
-from pydantic import field_validator, ConfigDict, BaseModel, FilePath, model_validator
 import logging
-import yaml
 from datetime import datetime
 from pathlib import Path
 from typing import Dict, List, Literal, Optional, Union
 
+import yaml
+from pydantic import BaseModel, ConfigDict, FilePath, field_validator, model_validator
 
 logger = logging.getLogger(__name__)
 
@@ -329,7 +329,7 @@ class Config(BaseModel):
             path.mkdir(parents=True)
         return path
 
-    @model_validator(mode='after')
+    @model_validator(mode="after")
     def check_date_order(self):
         if self.track_start_date > self.track_end_date:
             raise ValueError("track_end_date should be later than track_start_date")
