+++ src/wam2layers/preprocessing/shared.py
@@ -5,6 +5,8 @@ import pandas as pd
 import xarray as xr
 from scipy.interpolate import interp1d
 
+from wam2layers.tracking.backtrack import config
+
 
 # old, keep only for reference and ecearth starting case
 def resample(variable, divt, count_time, method="interp"):
@@ -282,3 +284,140 @@ def accumulation_to_flux(data):
     # Extrapolation introduces a small inconsistency at the last midnight...
     # data.interp(time=original_time, kwargs={"fill_value": "extrapolate"})
     return fluxdata
+
+
+def stagger_x(f):
+    """Interpolate f to the grid cell interfaces.
+
+    Only the values at the interior interfaces are returned
+
+    Arguments:
+        f: 2d array of shape [M, N]
+
+    Returns:
+        2d array of shape [M-2, N-1]
+    """
+    return 0.5 * (f[:, :-1] + f[:, 1:])[1:-1, :]
+
+
+def stagger_y(f):
+    """Interpolate f to the grid cell interfaces.
+
+    Only the values at the interior interfaces are returned
+
+    Arguments:
+        f: 2d array of shape [M, N]
+
+    Returns:
+        2d array of shape [M-1, N-2]
+    """
+    return 0.5 * (f[:-1, :] + f[1:, :])[:, 1:-1]
+
+
+def change_units(data, target_freq):
+    """Change units to m3.
+    Multiply by edge length or area to get flux in m3
+    Multiply by time to get accumulation instead of flux
+    Divide by density of water to go from kg to m3
+    """
+    density = 1000  # [kg/m3]
+    a, ly, lx = get_grid_info(data)
+
+    total_seconds = pd.Timedelta(target_freq).total_seconds()
+
+    for variable in data.data_vars:
+        if variable in ["fx_upper", "fx_lower"]:
+            data[variable] *= total_seconds / density * ly
+        elif variable in ["fy_upper", "fy_lower"]:
+            data[variable] *= total_seconds / density * lx[:, None]
+        elif variable in ["evap", "precip"]:
+            data[variable] *= total_seconds / density * a[:, None]
+        elif variable in ["s_upper", "s_lower"]:
+            data[variable] *= a[:, None] / density
+        else:
+            raise ValueError(f"Unrecognized variable {variable}")
+        data[variable] = data[variable].assign_attrs(units="m**3")
+
+
+def stabilize_fluxes(current, previous, progress_tracker, t):
+    """Stabilize the outfluxes / influxes.
+
+    CFL: Water cannot move further than one grid cell per timestep.
+    """
+    for level in ["upper", "lower"]:
+        fx = current["fx_" + level]
+        fy = current["fy_" + level]
+        s = previous["s_" + level]
+
+        fx_abs = np.abs(fx)
+        fy_abs = np.abs(fy)
+        ft_abs = fx_abs + fy_abs
+
+        fx_corrected = 1 / 2 * fx_abs / ft_abs * s.values
+        fx_stable = np.minimum(fx_abs, fx_corrected)
+
+        fy_corrected = 1 / 2 * fy_abs / ft_abs * s.values
+        fy_stable = np.minimum(fy_abs, fy_corrected)
+
+        progress_tracker.track_stability_correction(fy_corrected, fy_abs, config, t)
+
+        # Get rid of any nan values
+        fx_stable.fillna(0)
+        fy_stable.fillna(0)
+
+        # Re-instate the sign
+        current["fx_" + level] = np.sign(fx) * fx_stable
+        current["fy_" + level] = np.sign(fy) * fy_stable
+
+
+def convergence(fx, fy):
+    # Note: latitude decreasing, hence positive fy gradient is convergence
+    return np.gradient(fy, axis=-2) - np.gradient(fx, axis=-1)
+
+
+def calculate_fz(F, S0, S1):
+    """Calculate the vertical fluxes.
+
+    The vertical flux is calculated as a closure term. Residuals are distributed
+    proportionally to the amount of moisture in each layer.
+
+    The flux is constrained such that it can never exceed
+
+    Arguments:
+        F: xarray dataset with fluxes evaluated at temporal midpoints between states
+        S0: xarray dataset with states at current time t
+        S1: xarray dataset with states at updated time t+1 (always forward looking)
+
+    Returns:
+        fz: vertical flux, positive downward
+    """
+    s_mean = (S1 + S0) / 2
+    s_total = s_mean.s_upper + s_mean.s_lower
+    s_rel = s_mean / s_total
+
+    residual_upper = (
+        (S1 - S0).s_upper
+        - convergence(F.fx_upper, F.fy_upper)
+        + F.precip.values * s_rel.s_upper
+    )
+    residual_lower = (
+        (S1 - S0).s_lower
+        - convergence(F.fx_lower, F.fy_lower)
+        + F.precip.values * s_rel.s_lower
+        - F.evap
+    )
+
+    # compute the resulting vertical moisture flux; the vertical velocity so
+    # that the new residual_lower/s_lower = residual_upper/s_upper (positive downward)
+    fz = s_rel.s_lower * (residual_upper + residual_lower) - residual_lower
+
+    # stabilize the outfluxes / influxes; during the reduced timestep the
+    # vertical flux can maximally empty/fill 1/x of the top or down storage
+    stab = 1.0 / (config.kvf + 1.0)
+    flux_limit = np.minimum(
+        s_mean.s_upper, s_mean.s_lower
+    )  # TODO why is this not 'just' the upstream bucket?
+    fz_stable = np.minimum(np.abs(fz), stab * flux_limit)
+
+    # Reinstate the sign
+    return np.sign(fz) * fz_stable
