+++ src/wam2layers/tracking/backtrack.py
@@ -1,5 +1,4 @@
 import logging
-from functools import lru_cache, partial
 
 import click
 import numpy as np
@@ -7,8 +6,19 @@ import pandas as pd
 import xarray as xr
 
 from wam2layers.config import Config
-from wam2layers.preprocessing.shared import get_grid_info
-from wam2layers.utils import load_region
+from wam2layers.preprocessing.shared import (
+    calculate_fz,
+    change_units,
+    stabilize_fluxes,
+    stagger_x,
+    stagger_y,
+)
+from wam2layers.tracking.core import (
+    horizontal_advection,
+    vertical_advection,
+    vertical_dispersion,
+)
+from wam2layers.tracking.io import load_data, load_region, output_path
 from wam2layers.utils.profiling import ProgressTracker
 
 logger = logging.getLogger(__name__)
@@ -30,316 +40,6 @@ def get_tracking_dates(config):
         freq=config.target_frequency,
         inclusive="right",
     )
-
-
-def input_path(date, config):
-    input_dir = config.preprocessed_data_folder
-    return f"{input_dir}/{date.strftime('%Y-%m-%d')}_fluxes_storages.nc"
-
-
-def output_path(date, config):
-    output_dir = config.output_folder
-    return f"{output_dir}/backtrack_{date.strftime('%Y-%m-%dT%H-%M')}.nc"
-
-
-# LRU Cache keeps the file open so we save a bit on I/O
-@lru_cache(maxsize=2)
-def read_data_at_date(d):
-    """Load input data for given date."""
-    file = input_path(d, config)
-    return xr.open_dataset(file, cache=False)
-
-
-# This one can't be cached as we'll be overwriting the content.
-def read_data_at_time(t):
-    """Get a single time slice from input data at time t."""
-    ds = read_data_at_date(t)
-    return ds.sel(time=t, drop=True)
-
-
-def load_data(t, subset="fluxes"):
-    """Load variable at t, interpolate if needed."""
-    variables = {
-        "fluxes": ["fx_upper", "fx_lower", "fy_upper", "fy_lower", "evap", "precip"],
-        "states": ["s_upper", "s_lower"],
-    }
-
-    t1 = t.ceil(config.input_frequency)
-    da1 = read_data_at_time(t1)[variables[subset]]
-    if t == t1:
-        # this saves a lot of work if times are aligned with input
-        return da1
-
-    t0 = t.floor(config.input_frequency)
-    da0 = read_data_at_time(t0)[variables[subset]]
-    if t == t0:
-        return da0
-
-    return da0 + (t - t0) / (t1 - t0) * (da1 - da0)
-
-
-def load_region(config):
-    # TODO: make variable name more generic
-    return xr.open_dataset(config.region).source_region
-
-
-def stagger_x(f):
-    """Interpolate f to the grid cell interfaces.
-
-    Only the values at the interior interfaces are returned
-
-    Arguments:
-        f: 2d array of shape [M, N]
-
-    Returns:
-        2d array of shape [M-2, N-1]
-    """
-    return 0.5 * (f[:, :-1] + f[:, 1:])[1:-1, :]
-
-
-def stagger_y(f):
-    """Interpolate f to the grid cell interfaces.
-
-    Only the values at the interior interfaces are returned
-
-    Arguments:
-        f: 2d array of shape [M, N]
-
-    Returns:
-        2d array of shape [M-1, N-2]
-    """
-    return 0.5 * (f[:-1, :] + f[1:, :])[:, 1:-1]
-
-
-def pad_boundaries(*args, periodic=False):
-    """Add boundary padding to all input arrays.
-
-    Arguments:
-        *args: Input arrays to which boundary padding will be added.
-        periodic: If True, apply periodic boundary conditions when padding in
-            the x-direction. If False, pad with zeros in the x and y directions.
-
-    Returns:
-        List of input arrays with boundary padding added.
-    """
-    zero_xy = partial(np.pad, pad_width=1, mode="constant", constant_values=0)
-    periodic_x = partial(np.pad, pad_width=((0, 0), (1, 1)), mode="wrap")
-    zero_y = partial(
-        np.pad, pad_width=((1, 1), (0, 0)), mode="constant", constant_values=0
-    )
-
-    if periodic:
-        return [periodic_x(zero_y(arg)) for arg in args]
-    return [zero_xy(arg) for arg in args]
-
-
-def horizontal_advection(s, u, v, periodic_x=False) -> np.ndarray:
-    """Calculate advection on a staggered grid using a donor cell scheme.
-
-    Boundaries are padded with 0 or periodic boundaries to maintain the original
-    grid size.
-
-    The advection equation reads `grad(u*s)` where u is the velocity vector and
-    s is any scalar. Using i for grid cell indices in the x direction and j for
-    the y-direction (increasing northwards), the donor cell scheme reads:
-
-    s(i-1/2) = s(i-1) if u(i-1/2) > 0
-               s(i+1) otherwise
-    s(j-1/2) = s(j-1) if v(j-1/2) > 0
-               s(j+1) otherwise
-
-    d(us)/dx = u(i-1/2)*s(i-1/2) - u(i+1/2)*s(i+1/2)
-    d(vs)/dy = v(j-1/2)*s(j-1/2) - v(j+1/2)*s(j+1/2)
-
-    adv(s) = d(us)/dx + d(vs)/dy
-
-    Arguments:
-        s: array of shape [M, N]. It is assumed that the grid dimensions are
-            [latitude, longitude] and latitude is in decreasing order.
-        u: array of shape = [M-2, N-1]
-        v: array of shape = [M-1, N-2]
-
-    Returns:
-        array of shape [M, N]
-
-    Examples:
-
-        Create a simple array:
-        >>> s = np.zeros((5, 5))
-        >>> s[2, 2] = 1
-        >>> u = np.ones((3, 4)) * .5
-        >>> v = np.ones((4, 3)) * .5
-        >>> s
-        array([[0., 0., 0., 0., 0.],
-               [0., 0., 0., 0., 0.],
-               [0., 0., 1., 0., 0.],
-               [0., 0., 0., 0., 0.],
-               [0., 0., 0., 0., 0.]])
-
-        Calculate the advection for a single time step (forward)
-        >>> horizontal_advection(s, u, v)
-        array([[ 0. ,  0. ,  0. ,  0. ,  0. ],
-               [ 0. ,  0. ,  0.5,  0. ,  0. ],
-               [ 0. ,  0. , -1. ,  0.5,  0. ],
-               [ 0. ,  0. ,  0. ,  0. ,  0. ],
-               [ 0. ,  0. ,  0. ,  0. ,  0. ]])
-
-        Backtracking can be done by reversing the velocities:
-        >>> horizontal_advection(s, -u, -v)
-        array([[ 0. ,  0. ,  0. ,  0. ,  0. ],
-               [ 0. ,  0. ,  0. ,  0. ,  0. ],
-               [ 0. ,  0.5, -1. ,  0. ,  0. ],
-               [ 0. ,  0. ,  0.5,  0. ,  0. ],
-               [ 0. ,  0. ,  0. ,  0. ,  0. ]])
-
-    """
-    # Pad boundaries with ghost cells
-    qp, up, vp = pad_boundaries(s, u, v, periodic=periodic_x)
-    # shapes: [M+2, N+2], [M, N+1], [M+1, N]
-
-    west = np.s_[:-1]
-    east = np.s_[1:]
-    south = np.s_[1:]
-    north = np.s_[:-1]
-    inner = np.s_[1:-1]
-
-    # Donor cell upwind scheme (2 directions seperately)
-    uq = np.where(up > 0, up * qp[inner, west], up * qp[inner, east])  # [M, N+1]
-
-    vq = np.where(vp > 0, vp * qp[south, inner], vp * qp[north, inner])  # [M, N+2]
-
-    adv_x = uq[:, west] - uq[:, east]  # [M, N]
-    adv_y = vq[south, :] - vq[north, :]  # [M, N]
-
-    return adv_x + adv_y  # [M, N]
-
-
-def vertical_advection(fv, s_lower, s_upper):
-    """Calculate 1d upwind advection of vertical flux.
-
-    Upwind advection with fv positive downwards, so:
-
-    fv * s = fv * s_upper if fv > 0
-           = fv * s_lower otherwise
-    """
-    return np.where(fv >= 0, fv * s_upper, fv * s_lower)
-
-
-def vertical_dispersion(fv, s_lower, s_upper):
-    """Calculate additional vertical mixing due to convective dispersion.
-
-    dispersion = kvf * |Fv| * dS/dz
-    """
-    return config.kvf * np.abs(fv) * (s_upper - s_lower)
-
-
-def change_units(data, target_freq):
-    """Change units to m3.
-    Multiply by edge length or area to get flux in m3
-    Multiply by time to get accumulation instead of flux
-    Divide by density of water to go from kg to m3
-    """
-    density = 1000  # [kg/m3]
-    a, ly, lx = get_grid_info(data)
-
-    total_seconds = pd.Timedelta(target_freq).total_seconds()
-
-    for variable in data.data_vars:
-        if variable in ["fx_upper", "fx_lower"]:
-            data[variable] *= total_seconds / density * ly
-        elif variable in ["fy_upper", "fy_lower"]:
-            data[variable] *= total_seconds / density * lx[:, None]
-        elif variable in ["evap", "precip"]:
-            data[variable] *= total_seconds / density * a[:, None]
-        elif variable in ["s_upper", "s_lower"]:
-            data[variable] *= a[:, None] / density
-        else:
-            raise ValueError(f"Unrecognized variable {variable}")
-        data[variable] = data[variable].assign_attrs(units="m**3")
-
-
-def stabilize_fluxes(current, previous, progress_tracker, t):
-    """Stabilize the outfluxes / influxes.
-
-    CFL: Water cannot move further than one grid cell per timestep.
-    """
-    for level in ["upper", "lower"]:
-        fx = current["fx_" + level]
-        fy = current["fy_" + level]
-        s = previous["s_" + level]
-
-        fx_abs = np.abs(fx)
-        fy_abs = np.abs(fy)
-        ft_abs = fx_abs + fy_abs
-
-        fx_corrected = 1 / 2 * fx_abs / ft_abs * s.values
-        fx_stable = np.minimum(fx_abs, fx_corrected)
-
-        fy_corrected = 1 / 2 * fy_abs / ft_abs * s.values
-        fy_stable = np.minimum(fy_abs, fy_corrected)
-
-        progress_tracker.track_stability_correction(fy_corrected, fy_abs, config, t)
-
-        # Get rid of any nan values
-        fx_stable.fillna(0)
-        fy_stable.fillna(0)
-
-        # Re-instate the sign
-        current["fx_" + level] = np.sign(fx) * fx_stable
-        current["fy_" + level] = np.sign(fy) * fy_stable
-
-
-def convergence(fx, fy):
-    # Note: latitude decreasing, hence positive fy gradient is convergence
-    return np.gradient(fy, axis=-2) - np.gradient(fx, axis=-1)
-
-
-def calculate_fz(F, S0, S1):
-    """Calculate the vertical fluxes.
-
-    The vertical flux is calculated as a closure term. Residuals are distributed
-    proportionally to the amount of moisture in each layer.
-
-    The flux is constrained such that it can never exceed
-
-    Arguments:
-        F: xarray dataset with fluxes evaluated at temporal midpoints between states
-        S0: xarray dataset with states at current time t
-        S1: xarray dataset with states at updated time t+1 (always forward looking)
-
-    Returns:
-        fz: vertical flux, positive downward
-    """
-    s_mean = (S1 + S0) / 2
-    s_total = s_mean.s_upper + s_mean.s_lower
-    s_rel = s_mean / s_total
-
-    residual_upper = (
-        (S1 - S0).s_upper
-        - convergence(F.fx_upper, F.fy_upper)
-        + F.precip.values * s_rel.s_upper
-    )
-    residual_lower = (
-        (S1 - S0).s_lower
-        - convergence(F.fx_lower, F.fy_lower)
-        + F.precip.values * s_rel.s_lower
-        - F.evap
-    )
-
-    # compute the resulting vertical moisture flux; the vertical velocity so
-    # that the new residual_lower/s_lower = residual_upper/s_upper (positive downward)
-    fz = s_rel.s_lower * (residual_upper + residual_lower) - residual_lower
-
-    # stabilize the outfluxes / influxes; during the reduced timestep the
-    # vertical flux can maximally empty/fill 1/x of the top or down storage
-    stab = 1.0 / (config.kvf + 1.0)
-    flux_limit = np.minimum(
-        s_mean.s_upper, s_mean.s_lower
-    )  # TODO why is this not 'just' the upstream bucket?
-    fz_stable = np.minimum(np.abs(fz), stab * flux_limit)
-
-    # Reinstate the sign
-    return np.sign(fz) * fz_stable
 
 
 def backtrack(
